moved firebase init file out of next api
